+++ src/components/meta/Configure.aat.tsx
@@ -106,6 +106,17 @@ describe('The configure controls', () => {
     await userEvent.fill(input, 'Moorefly')
     expect(storage.config.PLAYER_NAME).toBe('Moorefly')
   })
+
+  it('should toggle the verbose setting', async () => {
+    const verboseCbx = screen.getByLabelText('verbose')
+    expect(storage.config.VERBOSE_BTN).toBe(true)
+
+    await verboseCbx.click()
+    expect(storage.config.VERBOSE_BTN).toBe(false)
+
+    await verboseCbx.click()
+    expect(storage.config.VERBOSE_BTN).toBe(true)
+  })
 })
 
 describe('The configure-page reset button', () => {
@@ -114,7 +125,7 @@ describe('The configure-page reset button', () => {
     storage.game = null
     storage.config = microConfig
     const screen = await renderWithProvider(<Configure/>)
-    await screen.getByTitle('Revert to Defaults').click()
+    await screen.getByTitle('Use Defaults').click()
     expect(storage.config).toStrictEqual(DEFAULTS)
   })